+++ src/main/java/com/tasksprints/auction/api/bid/BidController.java
@@ -5,8 +5,8 @@ import com.tasksprints.auction.common.response.ApiResult;
 import com.tasksprints.auction.domain.bid.dto.BidRequest;
 import com.tasksprints.auction.domain.bid.dto.BidResponse;
 import com.tasksprints.auction.domain.bid.service.BidService;
-import com.tasksprints.auction.domain.socket.dto.AddChatRoomDto;
-import com.tasksprints.auction.domain.socket.service.ChatService;
+import com.tasksprints.auction.domain.chat.dto.AddChatRoomDto;
+import com.tasksprints.auction.domain.chat.service.ChatService;
 import com.tasksprints.auction.domain.user.dto.response.UserDetailResponse;
 import com.tasksprints.auction.domain.user.model.User;
 import com.tasksprints.auction.domain.user.service.UserService;
